requirements.txt
flask
flask-pymongo
flask-mail
flask-cors
python-dotenv
PyJWT
bcrypt